Update package name and license; Upgrade phpunit version in composer
The package name in composer.json is updated to 'influactive/forms' from 'vendor_name/influactive-forms' for more accurate representation of the owned package. The license has been changed to GPL-2.0 from proprietary to respect open source license standards.

Additionally, the phpunit version has been upgraded from '10.2.7' to '10.3.1' in composer.lock file to ensure compatibility with the newest standards and security patches. Update has been reflected accordingly in the content hash, the dist url and reference, the branch alias, and the sources in support.