Update PHPCSExtra version and allow phpcodesniffer-composer-installer plugin
The update includes moving from `phpcsstandards/phpcsextra` version `1.1.0` to `1.1.1`. The revisions made in the version necessitated an update for better compatibility and functionality. Additionally, the `dealerdirect/phpcodesniffer-composer-installer` plugin is now allowed in the `composer.json`. This plugin helps with the installation and configuration of PHP CodeSniffer standards thus improving maintainability and readability of the codebase.