Refactor require_once syntax; Allow JS to load in footer
Changed the syntax of the 'require_once' inclusions, removing the parentheses for a simplified and cleaner look. This change does not affect the functionality of the code, but aligns it with common PHP coding standards.

Also, added 'true' as the last argument of the wp_enqueue_script function call. This instructs WordPress to load the JavaScript in the footer, helping to improve front-end loading time and performance as scripts will be loaded after the DOM is ready.